feat: disable s3 sync api route and remove button (#1277)
